Rename _cimax to cimax
